+++ src/Equinor.ProCoSys.Completion.WebApi/Controllers/MailTemplates/MailTemplatesController.cs
@@ -3,16 +3,17 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Threading;
 using System.Threading.Tasks;
-using Azure.Core;
 using Equinor.ProCoSys.Auth;
 using Equinor.ProCoSys.Common;
 using Equinor.ProCoSys.Common.Email;
 using Equinor.ProCoSys.Completion.Query.MailTemplateQueries.GetAllMailTemplates;
+using Equinor.ProCoSys.Completion.TieImport.CommonLib;
 using Equinor.ProCoSys.Completion.WebApi.Middleware;
 using Equinor.TI.CommonLibrary.Mapper.Core;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Options;
 
 namespace Equinor.ProCoSys.Completion.WebApi.Controllers.MailTemplates;
 
@@ -23,13 +24,15 @@ public class MailTemplatesController : ControllerBase
 {
     private readonly IMediator _mediator;
     private readonly IEmailService _emailService;
-    private readonly ISchemaSource _schemaSource;
+    //private readonly ISchemaSource _schemaSource;
+    private readonly IServiceProvider _serviceProvider;
 
     public MailTemplatesController(IMediator mediator, IEmailService emailService, IServiceProvider serviceProvide)
     {
         _mediator = mediator;
         _emailService = emailService;
-        _schemaSource = serviceProvide.GetRequiredKeyedService<ISchemaSource>("CommonLibApiSource");
+        _serviceProvider = serviceProvide;
+        //        _schemaSource = serviceProvide.GetRequiredKeyedService<ISchemaSource>("CommonLibApiSource");
     }
 
     /// <summary>
@@ -44,17 +47,18 @@ public class MailTemplatesController : ControllerBase
         return Ok(result);
     }
 
-    [HttpGet("TestSendMail-Delete-After-Test")]
-    public void SendTestMail(
+    [HttpGet("Test_CommonLibApiSource-Delete-After-Test")]
+    public void TestCommonLibApiSource(
         [FromHeader(Name = CurrentPlantMiddleware.PlantHeader)]
         [Required]
         [StringLength(PlantEntityBase.PlantLengthMax, MinimumLength = PlantEntityBase.PlantLengthMin)]
         string plant
         , CancellationToken cancellationToken)
     {
-        //await _emailService.SendEmailsAsync(["eha@equinor.com"], "subject", "Body", cancellationToken);
-        var schema = _schemaSource.Get("Completion", "NotCompletion");
-        Console.WriteLine(schema);
+        var schemaSource = _serviceProvider.GetRequiredKeyedService<ISchemaSource>("CommonLibApiSource");
+        var options = _serviceProvider.GetRequiredService<IOptions<CommonLibOptions>>();
+        var schema = schemaSource.Get(options.Value.SchemaFrom[0], options.Value.SchemaTo);
 
+        Console.WriteLine(schema);
     }
 }
