+++ src/Equinor.ProCoSys.Completion.WebApi/appsettings.json
@@ -51,7 +51,7 @@
     "CacheDurationDays": 1,
     "SchemaFrom": [ "Handover" ],
     "SchemaTo": "ProCoSys_TR3111",
-    "Scope": "https://StatoilSRM.onmicrosoft.com/Statoil.TI.CommonLibrary.UI/.default"
+    "Scope": ""
   },
   "ConnectionStrings": {
     "CompletionContext": "",
@@ -62,7 +62,15 @@
     "MailUserOid": "70f02efd-ba53-498c-8a75-09d2d7f31d61"
   },
   "Graph": {
-    "ClientId": "0ae66748-e5b8-418b-8f4f-291c368312df"
+    "TenantId": "3aa4a235-b6e2-48d5-9195-7fcf05b459b0",
+    "ClientId": "0ae66748-e5b8-418b-8f4f-291c368312df",
+    "ClientCredentials": [
+      {
+        "SourceType": "",
+        "KeyVaultUrl": "",
+        "KeyVaultCertificateName": ""
+      }
+    ]
   },
   "Logging": {
     "LogLevel": {
@@ -86,7 +94,7 @@
   },
   "MainApi": {
     "ApiVersion": "4.1",
-    "BaseAddress": "https://pcs-main-api-test.azurewebsites.net/api/"
+    "BaseAddress": "https://pcs-main-api-dev-pr.azurewebsites.net/api/"
   },
   "MigrateDatabase": false,
   "SeedDummyData": false,
@@ -102,15 +110,15 @@
     "Scopes": {}
   },
   "TieImport": {
-    "AdapterTieUri": "https://qa-tie-proxy.equinor.com",
-    "AzureCertificateName": "pcs-completion-test-tie-cert",
-    "AzureKeyVaultUrl": "https://kv-pcs5-completion-test.vault.azure.net/",
-    "AzureTieApiId": "246de5ab-6c09-4df7-aaab-370df915deea",
+    "AdapterTieUri": "",
+    "AzureCertificateName": "",
+    "AzureKeyVaultUrl": "",
+    "AzureTieApiId": "",
     "AdapterParallelMessageHandling": true,
     "AdapterMessageChunkSize": 40,
-    "AdapterIdleTimeBetweenBatch": 5,
-    "AdapterIdleTimeOnNoMessages": 50,
-    "AdapterSites": "ASGB,ASGC,ASGS",
+    "AdapterIdleTimeBetweenBatch": 0,
+    "AdapterIdleTimeOnNoMessages": 5000,
+    "AdapterSites": "",
     "AdapterApplication": "PROCOSYS5",
     "TestEnableTestFileMessageListener": false,
     "TestFileMessageListenerPath": "",
@@ -123,6 +131,6 @@
   },
   "UseAzureAppConfiguration": false,
   "Redis": {
-    "HostName": "redis-pcs5-completion-dev.redis.cache.windows.net:6380"
+    "HostName": ""
   }
 }
